+++ lib/alias.js
@@ -1,7 +1,7 @@
 "use strict";
 
 import inherit from "inherits";
-import Facade, { call } from "./facade";
+import { Facade } from "./facade";
 
 /**
  * Initialize a new `Alias` facade with a `dictionary` of arguments.
@@ -13,8 +13,8 @@ import Facade, { call } from "./facade";
  *
  * @augments Facade
  */
-function Alias(dictionary, opts) {
-  call(this, dictionary, opts);
+export function Alias(dictionary, opts) {
+  Facade.call(this, dictionary, opts);
 }
 
 inherit(Alias, Facade);
@@ -75,5 +75,3 @@ Alias.prototype.userId = function () {
  * @return {string}
  */
 Alias.prototype.to = Alias.prototype.userId;
-
-export default Alias;
