+++ lib/delete.js
@@ -1,7 +1,7 @@
 "use strict";
 
 import inherit from "inherits";
-import Facade, { call } from "./facade";
+import { Facade } from "./facade";
 
 /**
  * Initialize a new `Delete` facade with a `dictionary` of arguments.
@@ -14,8 +14,8 @@ import Facade, { call } from "./facade";
  *
  * @augments Facade
  */
-function Delete(dictionary, opts) {
-  call(this, dictionary, opts);
+export function Delete(dictionary, opts) {
+  Facade.call(this, dictionary, opts);
 }
 
 inherit(Delete, Facade);
@@ -28,5 +28,3 @@ inherit(Delete, Facade);
 Delete.prototype.type = function () {
   return "delete";
 };
-
-export default Delete;
