+++ lib/facade.js
@@ -34,7 +34,7 @@ import type from "type-component";
  * @see {@link https://github.com/segmentio/new-date|new-date}
  * @see {@link https://github.com/segmentio/isodate-traverse|isodate-traverse}
  */
-function Facade(obj, opts) {
+export function Facade(obj, opts) {
   opts = opts || {};
   if (!("clone" in opts)) opts.clone = true;
   if (opts.clone) obj = clone(obj);
@@ -546,5 +546,3 @@ address(Facade.prototype);
 function transform(obj) {
   return clone(obj);
 }
-
-export default Facade;
