+++ lib/group.js
@@ -3,7 +3,7 @@
 import inherit from "inherits";
 import isEmail from "is-email";
 import newDate from "new-date";
-import Facade from "./facade";
+import { Facade } from "./facade";
 
 /**
  * Initialize a new `Group` facade with a `dictionary` of arguments.
@@ -16,7 +16,7 @@ import Facade from "./facade";
  *
  * @augments Facade
  */
-function Group(dictionary, opts) {
+export function Group(dictionary, opts) {
   Facade.call(this, dictionary, opts);
 }
 
@@ -163,5 +163,3 @@ Group.prototype.properties = function () {
   // TODO remove this function
   return this.field("traits") || this.field("properties") || {};
 };
-
-export default Group;
