+++ lib/identify.js
@@ -1,6 +1,6 @@
 "use strict";
 
-import Facade from "./facade";
+import { Facade } from "./facade";
 import get from "obj-case";
 import inherit from "inherits";
 import isEmail from "is-email";
@@ -19,7 +19,7 @@ import type from "type-component";
  *
  * @augments Facade
  */
-function Identify(dictionary, opts) {
+export function Identify(dictionary, opts) {
   Facade.call(this, dictionary, opts);
 }
 
@@ -380,5 +380,3 @@ Identify.prototype.gender = Facade.proxy("traits.gender");
  * @return {object}
  */
 Identify.prototype.birthday = Facade.proxy("traits.birthday");
-
-export default Identify;
