+++ lib/index.js
@@ -1,21 +1,12 @@
 "use strict";
 
-import Facade, {
-  Alias,
-  Group,
-  Identify,
-  Track,
-  Page,
-  Screen,
-  Delete,
-} from "./facade";
+import { Facade as Fac } from "./facade";
 
-Alias = require("./alias").default;
-Group = require("./group").default;
-Identify = require("./identify").default;
-Track = require("./track").default;
-Page = require("./page").default;
-Screen = require("./screen").default;
-Delete = require("./delete");
-
-export default Facade;
+export { Alias } from "./alias";
+export { Group } from "./group";
+export { Identify } from "./identify";
+export { Track } from "./track";
+export { Page } from "./page";
+export { Screen } from "./screen";
+export { Delete } from "./delete";
+export const Facade = Fac;
