+++ lib/page.js
@@ -1,8 +1,8 @@
 "use strict";
 
 import inherit from "inherits";
-import Facade from "./facade";
-import Track from "./track";
+import { Facade } from "./facade";
+import { Track } from "./track";
 import isEmail from "is-email";
 
 /**
@@ -16,7 +16,7 @@ import isEmail from "is-email";
  *
  * @augments Facade
  */
-function Page(dictionary, opts) {
+export function Page(dictionary, opts) {
   Facade.call(this, dictionary, opts);
 }
 
@@ -208,5 +208,3 @@ Page.prototype.track = function (name) {
   json.properties = this.properties();
   return new Track(json, this.opts);
 };
-
-export default Page;
