+++ lib/screen.js
@@ -1,8 +1,8 @@
 "use strict";
 
 import inherit from "inherits";
-import Page from "./page";
-import Track from "./track";
+import { Page } from "./page";
+import { Track } from "./track";
 
 /**
  * Initialize a new `Screen` facade with a `dictionary` of arguments.
@@ -18,7 +18,7 @@ import Track from "./track";
  *
  * @augments Page
  */
-function Screen(dictionary, opts) {
+export function Screen(dictionary, opts) {
   Page.call(this, dictionary, opts);
 }
 
@@ -66,5 +66,3 @@ Screen.prototype.track = function (name) {
   json.properties = this.properties();
   return new Track(json, this.opts);
 };
-
-export default Screen;
