+++ lib/track.js
@@ -2,8 +2,8 @@
 
 import inherit from "inherits";
 import type from "type-component";
-import Facade from "./facade";
-import Identify from "./identify";
+import { Facade } from "./facade";
+import { Identify } from "./identify";
 import isEmail from "is-email";
 import get from "obj-case";
 
@@ -19,7 +19,7 @@ import get from "obj-case";
  *
  * @augments Facade
  */
-function Track(dictionary, opts) {
+export function Track(dictionary, opts) {
   Facade.call(this, dictionary, opts);
 }
 
@@ -612,5 +612,3 @@ function currency(val) {
     return val;
   }
 }
-
-export default Track;
