+++ package.json
@@ -33,6 +33,7 @@
     "@segment/eslint-config": "^3.1.1",
     "@segment/isodate": "^1.0.2",
     "@segment/to-iso-string": "^1.0.0",
+    "@types/dlv": "^1.1.2",
     "@types/jest": "^26.0.14",
     "browserify": "13.0.0",
     "browserify-istanbul": "^2.0.0",
@@ -53,6 +54,7 @@
     "karma-phantomjs-launcher": "^1.0.0",
     "karma-sauce-launcher": "^1.0.0",
     "karma-spec-reporter": "0.0.26",
+    "klona": "^2.0.4",
     "matcha": "^0.5.0",
     "mocha": "^2.4.5",
     "mochify": "^2.17.0",
